suspend pointer lock on overlay opened, shared takeFocus helper
