+++ src/services/overlay.ts
@@ -1,5 +1,7 @@
 import { IFRAME_MESSAGE_TYPE } from "@wvdsh/api";
 import { type WavedashSDK } from "../index";
+import { takeFocus } from "../utils/focus";
+import { suspendPointerLock } from "../utils/pointerLock";
 import { WavedashManager } from "./manager";
 
 /**
@@ -8,22 +10,40 @@ import { WavedashManager } from "./manager";
  * Owns the iframe ↔ parent interactions for the Wavedash overlay UI:
  * - Shift+Tab inside the iframe toggles the overlay on the host page
  *   (the host owns the overlay, so we postMessage up).
- * - When the parent closes the overlay it sends TAKE_FOCUS so keyboard
- *   input goes back to the game; we walk the DOM for a focusable target.
- * - `takeFocus()` is also called after load completes so the game starts
- *   with keyboard focus without the player clicking first.
+ * - When the parent closes the overlay it sends TAKE_FOCUS, which hands
+ *   keyboard focus back to the game (see `takeFocus`).
+ * - While the overlay is open we suspend pointer lock (the host broadcasts
+ *   OVERLAY_CHANGED) so a game can't hold/re-grab the cursor behind it.
  */
 export class OverlayManager extends WavedashManager {
+  // Restores native pointer lock; set while the overlay is open.
+  private restorePointerLock: (() => void) | undefined;
+
   constructor(sdk: WavedashSDK) {
     super(sdk);
 
     this.sdk.iframeMessenger.addEventListener(
       IFRAME_MESSAGE_TYPE.TAKE_FOCUS,
-      () => this.takeFocus()
+      () => takeFocus()
     );
+
+    /** @TODO uncomment once @wvdsh/api is published with latest IFRAME_MESSAGE_TYPE */
+    // this.sdk.iframeMessenger.addEventListener(
+    //   IFRAME_MESSAGE_TYPE.OVERLAY_CHANGED,
+    //   ({ isOpen }) => this.setOpen(isOpen)
+    // );
 
     if (typeof window !== "undefined") {
       window.addEventListener("keydown", this.handleKeyDown);
+    }
+  }
+
+  private setOpen(open: boolean): void {
+    if (open) {
+      this.restorePointerLock ??= suspendPointerLock();
+    } else {
+      this.restorePointerLock?.();
+      this.restorePointerLock = undefined;
     }
   }
 
@@ -32,23 +52,6 @@ export class OverlayManager extends WavedashManager {
       IFRAME_MESSAGE_TYPE.TOGGLE_OVERLAY,
       {}
     );
-  }
-
-  takeFocus(): void {
-    if (typeof document === "undefined") return;
-
-    const gameFocusTargets =
-      document.getElementsByClassName("game-focus-target");
-    if (gameFocusTargets.length > 0) {
-      (gameFocusTargets[0] as HTMLElement).focus();
-      return;
-    }
-
-    // Fallback: focus the first focusable element (canvas, input, button, etc.)
-    const focusableElement = document.querySelector(
-      "canvas, input, button, [tabindex]:not([tabindex='-1'])"
-    ) as HTMLElement | null;
-    focusableElement?.focus();
   }
 
   private handleKeyDown = (event: KeyboardEvent) => {
